+++ draft-packages/avatar/docs/Avatar.stories.tsx
@@ -16,6 +16,8 @@ export default {
 
 export const InitialsPersonal = () => (
   <>
+    <Avatar size="xlarge" fullName="John Smith" />
+    <br />
     <Avatar size="large" fullName="John Smith" />
     <br />
     <Avatar size="medium" fullName="John Smith" />
@@ -26,6 +28,8 @@ export const InitialsPersonal = () => (
 
 export const InitialsOthers = () => (
   <>
+    <Avatar isCurrentUser={false} size="xlarge" fullName="Jane Doe" />
+    <br />
     <Avatar isCurrentUser={false} size="large" fullName="Jane Doe" />
     <br />
     <Avatar isCurrentUser={false} size="medium" fullName="Jane Doe" />
@@ -36,6 +40,13 @@ export const InitialsOthers = () => (
 
 export const DisabledInitials = () => (
   <>
+    <Avatar
+      isCurrentUser={false}
+      size="xlarge"
+      fullName="213146147"
+      disableInitials
+    />
+    <br />
     <Avatar
       isCurrentUser={false}
       size="large"
@@ -62,6 +73,12 @@ export const DisabledInitials = () => (
 export const PhotoPersonal = () => (
   <>
     <Avatar
+      size="xlarge"
+      avatarSrc="https://www.cultureampcom-preview-1.usw2.wp-dev-us.cultureamp-cdn.com/assets/slices/main/assets/public/media/chapters-card-1@2x.05e547444387f29f14df0b82634bf2b6.png"
+      fullName="Jane Doe"
+    />
+    <br />
+    <Avatar
       size="large"
       avatarSrc="https://www.cultureampcom-preview-1.usw2.wp-dev-us.cultureamp-cdn.com/assets/slices/main/assets/public/media/chapters-card-1@2x.05e547444387f29f14df0b82634bf2b6.png"
       fullName="Jane Doe"
@@ -82,6 +99,8 @@ export const PhotoPersonal = () => (
 )
 export const DefaultUser = () => (
   <>
+    <Avatar size="xlarge" avatarSrc="broken" fullName="Jane Doe" />
+    <br />
     <Avatar size="large" avatarSrc="broken" fullName="Jane Doe" />
     <br />
     <Avatar size="medium" avatarSrc="broken" fullName="Jane Doe" />
@@ -92,6 +111,8 @@ export const DefaultUser = () => (
 
 export const WithLongInitials = () => (
   <>
+    <Avatar size="xlarge" fullName="Very Long Name Which Shows Initials" />
+    <br />
     <Avatar size="large" fullName="Very Long Name Which Shows Initials" />
     <br />
     <Avatar size="medium" fullName="Very Long Name Which Shows Initials" />