+++ draft-packages/illustration/KaizenDraft/Illustration/Players/LottiePlayer.tsx
@@ -1,13 +1,19 @@
 import React, { useEffect, useRef, useState } from "react"
 import lottie from "lottie-web"
 import { assetUrl } from "@kaizen/hosted-assets"
-import { BaseProps } from "./Base"
-import { getAnimationData } from "./utils"
-import styles from "./style.module.scss"
-import { LottieAnimation } from "./types"
+import { getAnimationData } from "../utils"
+import styles from "../style.module.scss"
+import { LottieAnimation } from "../types"
+import { BaseProps } from "../Base"
 
 export type AnimatedBaseProps = {
+  /**
+   * Specifies whether the animation plays as soon as it is rendered
+   */
   autoplay?: boolean
+  /**
+   * Replay from start when active animation reaches the end of the animation
+   */
   loop?: boolean
 }