+++ draft-packages/empty-state/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.4](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-empty-state@2.3.3...@kaizen/draft-empty-state@2.3.4) (2021-07-02)
+
+**Note:** Version bump only for package @kaizen/draft-empty-state
+
+
+
+
+
+## [2.3.3](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-empty-state@2.3.2...@kaizen/draft-empty-state@2.3.3) (2021-06-30)
+
+**Note:** Version bump only for package @kaizen/draft-empty-state
+
+
+
+
+
+## [2.3.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-empty-state@2.3.1...@kaizen/draft-empty-state@2.3.2) (2021-06-25)
+
+**Note:** Version bump only for package @kaizen/draft-empty-state
+
+
+
+
+
+## [2.3.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-empty-state@2.3.0...@kaizen/draft-empty-state@2.3.1) (2021-06-23)
+
+**Note:** Version bump only for package @kaizen/draft-empty-state
+
+
+
+
+
 # [2.3.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-empty-state@2.2.3...@kaizen/draft-empty-state@2.3.0) (2021-06-11)