+++ draft-packages/form/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.6.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-form@3.6.1...@kaizen/draft-form@3.6.2) (2021-07-02)
+
+**Note:** Version bump only for package @kaizen/draft-form
+
+
+
+
+
+## [3.6.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-form@3.6.0...@kaizen/draft-form@3.6.1) (2021-07-01)
+
+**Note:** Version bump only for package @kaizen/draft-form
+
+
+
+
+
+# [3.6.0](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-form@3.5.1...@kaizen/draft-form@3.6.0) (2021-06-24)
+
+
+### Features
+
+* controlHasUnconstrainedHeight option for Elm Select ([#1715](https://github.com/cultureamp/kaizen-design-system/issues/1715)) ([374edce](https://github.com/cultureamp/kaizen-design-system/commit/374edce3c130153ec90b86a55b8eb6f581382e8c))
+
+
+
+
+
 ## [3.5.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-form@3.5.0...@kaizen/draft-form@3.5.1) (2021-06-11)
 
 **Note:** Version bump only for package @kaizen/draft-form