+++ draft-packages/guidance-block/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.1.9](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@3.1.8...@kaizen/draft-guidance-block@3.1.9) (2021-07-02)
+
+**Note:** Version bump only for package @kaizen/draft-guidance-block
+
+
+
+
+
+## [3.1.8](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@3.1.7...@kaizen/draft-guidance-block@3.1.8) (2021-06-25)
+
+**Note:** Version bump only for package @kaizen/draft-guidance-block
+
+
+
+
+
+## [3.1.7](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@3.1.6...@kaizen/draft-guidance-block@3.1.7) (2021-06-23)
+
+**Note:** Version bump only for package @kaizen/draft-guidance-block
+
+
+
+
+
 ## [3.1.6](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-guidance-block@3.1.5...@kaizen/draft-guidance-block@3.1.6) (2021-06-17)
 
 **Note:** Version bump only for package @kaizen/draft-guidance-block