+++ draft-packages/loading-placeholder/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.21](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-loading-placeholder@1.5.20...@kaizen/draft-loading-placeholder@1.5.21) (2021-07-02)
+
+**Note:** Version bump only for package @kaizen/draft-loading-placeholder
+
+
+
+
+
 ## [1.5.20](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-loading-placeholder@1.5.19...@kaizen/draft-loading-placeholder@1.5.20) (2021-06-11)
 
 **Note:** Version bump only for package @kaizen/draft-loading-placeholder