+++ draft-packages/modal/CHANGELOG.md
@@ -3,6 +3,46 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.6.6](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@5.6.5...@kaizen/draft-modal@5.6.6) (2021-07-02)
+
+**Note:** Version bump only for package @kaizen/draft-modal
+
+
+
+
+
+## [5.6.5](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@5.6.4...@kaizen/draft-modal@5.6.5) (2021-07-01)
+
+**Note:** Version bump only for package @kaizen/draft-modal
+
+
+
+
+
+## [5.6.4](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@5.6.3...@kaizen/draft-modal@5.6.4) (2021-06-25)
+
+**Note:** Version bump only for package @kaizen/draft-modal
+
+
+
+
+
+## [5.6.3](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@5.6.2...@kaizen/draft-modal@5.6.3) (2021-06-24)
+
+**Note:** Version bump only for package @kaizen/draft-modal
+
+
+
+
+
+## [5.6.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@5.6.1...@kaizen/draft-modal@5.6.2) (2021-06-23)
+
+**Note:** Version bump only for package @kaizen/draft-modal
+
+
+
+
+
 ## [5.6.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-modal@5.6.0...@kaizen/draft-modal@5.6.1) (2021-06-15)