+++ packages/stylelint-plugin/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.2](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/stylelint-plugin@1.1.1...@kaizen/stylelint-plugin@1.1.2) (2021-07-02)
+
+**Note:** Version bump only for package @kaizen/stylelint-plugin
+
+
+
+
+
+## [1.1.1](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/stylelint-plugin@1.1.0...@kaizen/stylelint-plugin@1.1.1) (2021-06-28)
+
+
+### Bug Fixes
+
+* use module.exports to fix stylelint plugin ([#1728](https://github.com/cultureamp/kaizen-design-system/issues/1728)) ([71ff851](https://github.com/cultureamp/kaizen-design-system/commit/71ff851b2d938135145d3406f540d903798fcaff))
+
+
+
+
+
 # 1.1.0 (2021-06-15)